delete access controll added

+++ REST API/src/controllers/Controller.js
@@ -39,8 +39,9 @@ class Controller {
     update = async (req, res) => {
         const input = req.body;
         const id = req.params.id;
+        const user = req.user;
         try {
-            const updatedEntity = await this.service.update(id, input);
+            const updatedEntity = await this.service.update(id, input, user);
             res.json(updatedEntity);
         } catch (error) {
             console.log(error);
@@ -49,14 +50,21 @@ class Controller {
     }
 
     delete = async (req, res) => {
-        const id = req.params.id;
-        try {
-            await this.service.deleteById(id);
-            res.status(204).json({});
-        } catch (error) {
-            console.log(error);
-            res.status(error.status || 400).json(errorHandler(error));
+        // Only admins can delete
+        if (req?.user.role === 'admin') {
+            const id = req.params.id;
+            try {
+                await this.service.deleteById(id);
+                res.status(204).json({});
+            } catch (error) {
+                console.log(error);
+                res.status(error.status || 400).json(errorHandler(error));
+            }
+        } else {
+            return res.status(401).json({ message: "Unauthorized" })
         }
+
+
     }
 }
